requirements.txt
# djangorestframework
# Pillow
# djangorestframework-simplejwt

asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.2
pillow==11.0.0
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.2
joblib
scikit-learn 
pandas

